BUG FIXES
fixed a ton of bugs like render overflow and function issues.

+++ lib/providers/workout_provider.dart
@@ -50,7 +50,7 @@ class WorkoutProvider with ChangeNotifier {
       Duration duration,
       WorkoutType type,
       List<Exercise> exercises,
-      double caloriesBurned,
+      int caloriesBurned,
       {String? notes}) async {
     _isLoading = true;
     notifyListeners();
